requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-tavily==0.2.11
langgraph==0.6.5
streamlit==1.48.1
notebook>=7.4.5
pyyaml